requirements-listen.txt
mlx-whisper>=0.4.0
openai-whisper>=20230314
faster-whisper>=1.0.0
sounddevice>=0.4.6
numpy>=1.24.0
scipy>=1.10.0
